backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
tensorflow>=2.13.0
librosa>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
soundfile>=0.12.0
audioread>=3.0.0
pillow>=10.0.0
python-decouple>=3.8